+++ jdaviz/configs/mosviz/plugins/viewers.py
@@ -43,7 +43,7 @@ class MosvizImageView(JdavizViewerMixin, BqplotImageView, AstrowidgetsImageViewe
         self.figure.fig_margin = {'left': 0, 'bottom': 0, 'top': 0, 'right': 0}
 
     def data(self, cls=None):
-        return [layer_state.layer  # .get_object(cls=cls or self.default_class)
+        return [layer_state.layer.get_object(cls=cls or self.default_class)
                 for layer_state in self.state.layers
                 if hasattr(layer_state, 'layer') and
                 isinstance(layer_state.layer, BaseData)]
@@ -207,8 +207,9 @@ class MosvizProfile2DView(JdavizViewerMixin, BqplotImageView):
             x_scales.min = max(limits) if self.inverted_x_axis else min(limits)
             x_scales.max = min(limits) if self.inverted_x_axis else max(limits)
 
-    def data(self, cls=None):
-        return [layer_state.layer.get_object(cls=cls or self.default_class)
+    def data(self, cls=None, statistic=None):
+        return [layer_state.layer.get_object(statistic=statistic,
+                                             cls=cls or self.default_class)
                 for layer_state in self.state.layers
                 if hasattr(layer_state, 'layer') and
                 isinstance(layer_state.layer, BaseData)]